+++ Code/Simple/Preallocated/Node.h
@@ -31,14 +31,14 @@ private:
 public:
     Node();
     Node(vector<unsigned int>* input, unsigned int alphabetMin, unsigned int alphabetMax, Node* parentNode,
-        Node* &node_pt, bitmap_t* in_bitmap, unsigned long &in_bitmapOffset, uint skew);
-    int rank(int character, unsigned long index, bitmap_t* bitmap, int alphabetMin, int alphabetMax, uint skew);
+        Node* &node_pt, bitmap_t* in_bitmap, unsigned long &in_bitmapOffset, float skew);
+    int rank(int character, unsigned long index, bitmap_t* bitmap, int alphabetMin, int alphabetMax, float skew);
     unsigned long int popcountBinaryRank(unsigned long pos, bitmap_t* bitmap);
     ulong binaryRank(ulong index, bitmap_t* bitmap);
     int select(int character, bool charBit, unsigned long occurance, bitmap_t* bitmap);
     int binarySelect(bool charBit, unsigned long occurance, bitmap_t* bitmap);
     int popcountBinarySelect(bool charBit, unsigned long occurance, bitmap_t* bitmap);
-    Node* getLeaf(int character, int alphabetMin, int alphabetMax, uint skew);
+    Node* getLeaf(int character, int alphabetMin, int alphabetMax, float skew);
     int leafSelect(int character, unsigned long occurance, bitmap_t* bitmap);
     
     inline void showBits(unsigned long n);
